splitting up thread logic
