+++ client/src/api/hooks/threads/index.ts
@@ -1,0 +1,4 @@
+export * from './useThreadData.js';
+export * from './useThreadActions.js';
+export * from './useThreadReplies.js';
+export * from './useThreadsList.js';
