+++ client/src/api/index.ts
@@ -4,11 +4,16 @@ export * from './queries/thread';
 export * from './queries/notification';
 export * from './queries/user';
 
-// Hook exports
+// Export general hooks
+export * from './hooks/useUserProfile';
 export * from './hooks/usePoll';
-export * from './hooks/useThreads';
-export * from './hooks/useThread';
-export * from './hooks/useCreatePost';
-export * from './hooks/useNotifications';
 export * from './hooks/useTopUsers';
-export * from './hooks/useUserProfile'; 
+export * from './hooks/useNotifications';
+export * from './hooks/useCreatePost';
+
+// Export thread-specific hooks directly
+export * from './hooks/threads';
+
+// Re-export with aliases for backward compatibility
+import { useThreadsList } from './hooks/threads';
+export { useThreadsList as useThreads };
